Issue 04: E2E tests on CharacterBaseInformation
