+++ package.json
@@ -15,7 +15,7 @@
   },
   "dependencies": {
     "core-js": "^3.6.4",
-    "tailwindcss": "^1.5.1",
+    "tailwindcss": "^1.5.2",
     "vue": "^2.6.11",
     "vuex": "^3.5.1"
   },
@@ -33,6 +33,7 @@
     "@vue/eslint-config-prettier": "^6.0.0",
     "@vue/eslint-config-typescript": "^5.0.2",
     "@vue/test-utils": "1.0.3",
+    "cypress-localstorage-commands": "^1.2.1",
     "eslint": "^7.5.0",
     "eslint-plugin-prettier": "^3.1.4",
     "eslint-plugin-vue": "^6.2.2",
