+++ src/Domain/CharacterBase/Repository/CharacterBaseInformationRepository.ts
@@ -1,0 +1,6 @@
+import { CharacterBaseInformation } from '@/Domain/CharacterBase/Types/CharacterBaseInformation';
+
+export default interface CharacterBaseInformationRepository {
+  get(): string | null;
+  save(characterBaseInformation: CharacterBaseInformation): void;
+}
