+++ src/Infrastructure/UI/Vue/App.vue
@@ -1,0 +1,19 @@
+<template>
+  <div id="app" class="container container-xl mx-auto">
+    <CharacterBase />
+  </div>
+</template>
+
+<script lang="ts">
+import Vue from 'vue';
+import CharacterBase from '@/Infrastructure/UI/Vue/components/CharacterBase/CharacterBase.vue';
+
+export default Vue.extend({
+  name: 'App',
+  components: {
+    CharacterBase,
+  },
+});
+</script>
+
+<style src="./assets/css/tailwind.css" />
