+++ src/Infrastructure/UI/Vue/assets/css/tailwind.css
