+++ src/Infrastructure/UI/Vue/assets/images/favicon.png
