+++ src/Infrastructure/UI/Vue/assets/images/logo.png
