--- src/domain/characterBase/repository/CharacterBaseInformationRepository.ts
@@ -1,6 +1,0 @@
-import { CharacterBaseInformation } from '@/domain/characterBase/types/CharacterBaseInformation';
-
-export default interface CharacterBaseInformationRepository {
-  get(): string | null;
-  save(characterBaseInformation: CharacterBaseInformation): void;
-}
