--- src/domain/characterBase/types/NewCharacterBaseValue.ts
@@ -1,8 +1,0 @@
-import { CharacterBaseInformation } from '@/domain/characterBase/types/CharacterBaseInformation';
-
-type NewCharacterBaseValue = {
-  identifier: keyof CharacterBaseInformation;
-  value: string;
-};
-
-export { NewCharacterBaseValue };
