+++ tests/e2e/plugins/index.js
@@ -16,10 +16,12 @@ module.exports = (on, config) => {
   // }))
 
   return Object.assign({}, config, {
-    fixturesFolder: 'tests/e2e/fixtures',
+    fixturesFolder: 'tests/fixtures',
     integrationFolder: 'tests/e2e/specs',
     screenshotsFolder: 'tests/e2e/screenshots',
     videosFolder: 'tests/e2e/videos',
     supportFile: 'tests/e2e/support/index.js',
+    viewportHeight: 800,
+    viewportWidth: 1200,
   });
 };
