--- tests/e2e/specs/test.js
@@ -1,8 +1,0 @@
-// https://docs.cypress.io/api/introduction/api.html
-
-describe('My First Test', () => {
-  it('Visits the app root url', () => {
-    cy.visit('/');
-    cy.contains('Nom du personnage');
-  });
-});
