+++ tests/e2e/support/commands.js
@@ -23,3 +23,5 @@
 //
 // -- This is will overwrite an existing command --
 // Cypress.Commands.overwrite("visit", (originalFn, url, options) => { ... })
+
+import 'cypress-localstorage-commands';
