--- tests/fixtures/CharacterBaseInformation.ts
@@ -1,19 +1,0 @@
-import { CharacterBaseInformation } from '@/domain/characterBase/types/CharacterBaseInformation';
-
-export const fixtures: CharacterBaseInformation = {
-  name: 'Damien',
-  class: 'coder',
-  level: '1',
-  sizeCategory: 'l',
-  size: '1 meter 85',
-  age: '38',
-  sex: 'm',
-  playerName: 'Carcel',
-  race: 'human',
-  alignment: 'cm',
-  divinity: 'none',
-  weight: '105 kg',
-  eyes: 'brown',
-  hair: 'brown',
-  skin: 'pale',
-};
