+++ tests/fixtures/CharacterBaseInformation/empty.json
@@ -1,0 +1,17 @@
+{
+  "name": "",
+  "class": "",
+  "level": "",
+  "sizeCategory": "",
+  "size": "",
+  "age": "",
+  "sex": "",
+  "playerName": "",
+  "race": "",
+  "alignment": "",
+  "divinity": "",
+  "weight": "",
+  "eyes": "",
+  "hair": "",
+  "skin": ""
+}
