--- tests/unit/infrastructure/ui/components/CharacterBase/FlatInput.unit.ts
@@ -1,5 +1,0 @@
-describe('Unit test example', () => {
-  it('tests nothing', () => {
-    expect(true).toStrictEqual(true);
-  });
-});
