+++ tsconfig.json
@@ -12,6 +12,7 @@
         "src/*"
       ]
     },
+    "resolveJsonModule": true,
     "sourceMap": true,
     "strict": true,
     "target": "esnext",
